prevent reload on todo submition
