+++ src/index.js
@@ -1,9 +1,8 @@
 import inputForms from './landing';
 import asideMod from './aside';
-import project from './project';
-import forms from './forms';
+import forms from './project';
 
-project();
+forms.project();
 inputForms.landing();
 forms.toDoForm();
 asideMod.aside();
